NetworkGraph 설정 함수 분리
