+++ src/constants/graphIconMap.js
@@ -1,0 +1,50 @@
+export const graphIconMap = new Map([
+  // 매출 관련
+  ["매출이익", "매출이익.png"],
+  ["매출액", "매출액.png"],
+  ["매출원가", "매출원가.png"],
+
+  // 제품군 / 재고
+  ["FERT100s", "와인박스.png"],
+  ["FERT200s", "와인박스.png"],
+  ["기초재고", "기초재고.png"],
+  ["기말재고", "기말재고.png"],
+  ["당기제품제조원가", "비용2.png"],
+
+  // 제품 단위
+  ["FERT101", "와인.png"],
+  ["FERT102", "와인.png"],
+  ["FERT103", "와인.png"],
+  ["FERT104", "와인.png"],
+  ["FERT105", "와인.png"],
+  ["FERT106", "와인.png"],
+  ["FERT201", "와인.png"],
+  ["FERT202", "와인.png"],
+  ["FERT203", "와인.png"],
+
+  // 제조비용
+  ["당기제조비용", "비용.png"],
+  ["재공품", "재공품.png"],
+  ["액티비티배부", "비용2.png"],
+  ["액티비티단수차", "비용2.png"],
+
+  // 세부 원가
+  ["원재료비", "비용.png"],
+  ["부재료비", "비용.png"],
+  ["가공비", "비용.png"],
+  ["생산입고", "생산입고.png"],
+  ["공정출고", "공정출고.png"],
+  ["액티비티단가합", "비용2.png"],
+  ["액티비티수차합", "비용2.png"],
+
+  // 포도 재료
+  ["ROH0001누적", "포도.png"],
+  ["ROH0002누적", "포도.png"],
+  ["ROH0003누적", "포도.png"],
+  ["ROH2001누적", "포도.png"],
+  ["ROH2002누적", "포도.png"],
+  ["ROH2003누적", "포도.png"],
+
+  // 비용 계획
+  ["비용계획합", "비용.png"],
+]);
